# packages/contracts/openapi/company-api.v1.yaml
openapi: 3.1.0
info:
  title: Company DSH API
  version: 1.0.0
  description: >-
    Frozen browser-facing contract for the three-worktree baseline. All paths are
    served below /company-api/v1. Mutations require the company_session cookie,
    X-CSRF-Token, and expected_version where the request schema declares it.
servers:
  - url: /company-api/v1
security:
  - cookieAuth: []
...
paths:
...
  /daily-reports:
    get:
      tags: [DailyReports]
      operationId: listDailyReports
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - name: from
          in: query
          schema: { $ref: '#/components/schemas/DateString' }
        - name: to
          in: query
          schema: { $ref: '#/components/schemas/DateString' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/DailyReportStatus' }
        - name: scope
          in: query
          schema: { $ref: '#/components/schemas/DailyReportScope' }
        - name: task_id
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
      responses:
        '200':
          description: Current user's reports
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DailyReportPage' }
        default: { $ref: '#/components/responses/Problem' }
  /daily-reports/{work_date}:
    parameters:
      - { $ref: '#/components/parameters/WorkDate' }
      - { $ref: '#/components/parameters/DailyReportScope' }
      - { $ref: '#/components/parameters/DailyReportTaskId' }
    get:
      tags: [DailyReports]
      operationId: getDailyReport
      responses:
        '200':
          description: Daily report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DailyReport' }
        default: { $ref: '#/components/responses/Problem' }
    put:
      tags: [DailyReports]
      operationId: upsertDailyReport
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertDailyReportRequest' }
      responses:
        '200':
          description: Daily report saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DailyReport' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      tags: [DailyReports]
      operationId: deleteDailyReport
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - name: expected_version
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '204': { description: Soft deleted }
        default: { $ref: '#/components/responses/Problem' }
  /daily-reports/{work_date}/publish:
    post:
      tags: [DailyReports]
      operationId: publishDailyReport
      parameters:
        - { $ref: '#/components/parameters/WorkDate' }
        - { $ref: '#/components/parameters/DailyReportScope' }
        - { $ref: '#/components/parameters/DailyReportTaskId' }
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionedCommand' }
      responses:
        '200':
          description: Published daily report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DailyReport' }
        default: { $ref: '#/components/responses/Problem' }
  /daily-reports/{work_date}/rewrite-runs:
    post:
      tags: [DailyReports]
      operationId: startDailyReportRewrite
      parameters:
        - { $ref: '#/components/parameters/WorkDate' }
        - { $ref: '#/components/parameters/DailyReportScope' }
        - { $ref: '#/components/parameters/DailyReportTaskId' }
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DailyRewriteRequest' }
      responses:
        '202':
          description: Rewrite accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutomationOperation' }
        default: { $ref: '#/components/responses/Problem' }
  /daily-reports/{work_date}/apply-rewrite:
    post:
      tags: [DailyReports]
      operationId: applyDailyReportRewrite
      parameters:
        - { $ref: '#/components/parameters/WorkDate' }
        - { $ref: '#/components/parameters/DailyReportScope' }
        - { $ref: '#/components/parameters/DailyReportTaskId' }
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplyDailyRewriteRequest' }
      responses:
        '200':
          description: Applied daily report rewrite
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DailyReport' }
        default: { $ref: '#/components/responses/Problem' }
  /departments/{department_id}/daily-reports:
    get:
      tags: [DailyReports]
      operationId: listDepartmentDailyReports
      parameters:
        - { $ref: '#/components/parameters/DepartmentId' }
        - name: date
          in: query
          schema: { $ref: '#/components/schemas/DateString' }
        - name: from
          in: query
          schema: { $ref: '#/components/schemas/DateString' }
        - name: to
          in: query
          schema: { $ref: '#/components/schemas/DateString' }
        - name: member_user_id
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/DailyReportStatus' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Department reports and missing members
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentDailyReportView' }
        default: { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: company_session
  parameters:
    CsrfToken: { name: X-CSRF-Token, in: header, required: true, schema: { type: string } }
    IdempotencyKey:
      {
        name: Idempotency-Key,
        in: header,
        required: true,
        schema: { type: string, minLength: 8, maxLength: 128 },
      }
    Cursor: { name: cursor, in: query, schema: { type: [string, 'null'] } }
    Limit:
      { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
...
    DepartmentId:
      {
        name: department_id,
        in: path,
        required: true,
        schema: { $ref: '#/components/schemas/Uuid' },
      }
...
    WorkDate:
      {
        name: work_date,
        in: path,
        required: true,
        schema: { $ref: '#/components/schemas/DateString' },
      }
    DailyReportScope:
      {
        name: scope,
        in: query,
        required: true,
        schema: { $ref: '#/components/schemas/DailyReportScope' },
      }
    DailyReportTaskId: { name: task_id, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
...
  responses:
    Problem:
      description: RFC 9457 compatible problem response
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
    DateString: { type: string, format: date, pattern: '^\d{4}-\d{2}-\d{2}$' }
    PlatformRole: { type: string, enum: [admin, member] }
...
    ReviewResult: { type: string, enum: [pass, fail, needs_review] }
    AutomationRunStatus: { type: string, enum: [queued, running, succeeded, failed, cancelled] }
    DailyReportStatus: { type: string, enum: [draft, published, deleted] }
    DailyReportScope: { type: string, enum: [personal, department, company, task] }
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, code, request_id]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        request_id: { type: string }
        field_errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    UserSummary:
      type: object
      required: [id, username, display_name, platform_role]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        username: { type: string }
        display_name: { type: string }
        platform_role: { $ref: '#/components/schemas/PlatformRole' }
    Department:
      type: object
      required: [id, name, status, version]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        name: { type: string }
        status: { type: string, enum: [active, disabled] }
        version: { type: integer, minimum: 1 }
...
    VersionedCommand:
      type: object
      additionalProperties: false
      required: [expected_version]
      properties:
        expected_version: { type: integer, minimum: 1 }
...
    SplitTaskDraft:
      type: object
      required:
        [
          client_id,
          title,
          description,
          acceptance_criteria,
          assignee_user_id,
          depends_on_client_ids,
          position,
        ]
      properties:
        client_id: { type: string }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 20000 }
        acceptance_criteria: { type: array, items: { type: string } }
        assignee_user_id:
          oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }]
        depends_on_client_ids: { type: array, items: { type: string } }
        position: { type: integer, minimum: 0 }
...
    EvidenceItem:
      type: object
      required: [kind, label, value]
      properties:
        kind: { type: string, enum: [url, text, file_ref] }
        label: { type: string, maxLength: 200 }
        value: { type: string, maxLength: 10000 }
...
    ReviewCheck:
      type: object
      required: [name, passed, detail]
      properties:
        name: { type: string }
        passed: { type: [boolean, 'null'] }
        detail: { type: string }
...
    AutomationOperation:
      type: object
      required: [id, kind, status, result, error, created_at, completed_at]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        kind: { type: string, enum: [requirement_split, task_review, daily_rewrite] }
        status: { $ref: '#/components/schemas/AutomationRunStatus' }
        result:
          oneOf:
            - { $ref: '#/components/schemas/RequirementSplitResult' }
            - { $ref: '#/components/schemas/TaskReviewAutomationResult' }
            - { $ref: '#/components/schemas/DailyRewriteResult' }
            - type: 'null'
        error:
          oneOf:
            - type: object
              required: [code, message]
              properties: { code: { type: string }, message: { type: string } }
            - type: 'null'
        created_at: { $ref: '#/components/schemas/Timestamp' }
        completed_at:
          oneOf: [{ $ref: '#/components/schemas/Timestamp' }, { type: 'null' }]
    RequirementSplitResult:
      type: object
      additionalProperties: false
      required: [tasks]
      properties:
        tasks:
          {
            type: array,
            items: { $ref: '#/components/schemas/SplitTaskDraft' },
            minItems: 1,
            maxItems: 100,
          }
    TaskReviewAutomationResult:
      type: object
      additionalProperties: false
      required: [result, summary, checks, evidence, executor_version]
      properties:
        result: { $ref: '#/components/schemas/ReviewResult' }
        summary: { type: string }
        checks: { type: array, items: { $ref: '#/components/schemas/ReviewCheck' } }
        evidence: { type: array, items: { $ref: '#/components/schemas/EvidenceItem' } }
        executor_version: { type: string }
    DailyRewriteResult:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content: { $ref: '#/components/schemas/DailyReportContent' }
    DailyReportContent:
      type: object
      required: [completed_today, next_plan, blockers, other, free_text]
      properties:
        completed_today: { type: string, maxLength: 20000 }
        next_plan: { type: string, maxLength: 20000 }
        blockers: { type: string, maxLength: 20000 }
        other: { type: string, maxLength: 20000 }
        free_text: { type: [string, 'null'], maxLength: 20000 }
    DailyReport:
      type: object
      required:
        [
          id,
          user_id,
          department_id,
          work_date,
          scope,
          task_id,
          content,
          status,
          version,
          published_at,
          updated_at,
        ]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        user_id: { $ref: '#/components/schemas/Uuid' }
        department_id: { $ref: '#/components/schemas/Uuid' }
        work_date: { $ref: '#/components/schemas/DateString' }
        scope: { $ref: '#/components/schemas/DailyReportScope' }
        task_id:
          oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }]
        content: { $ref: '#/components/schemas/DailyReportContent' }
        status: { $ref: '#/components/schemas/DailyReportStatus' }
        version: { type: integer, minimum: 1 }
        published_at:
          oneOf: [{ $ref: '#/components/schemas/Timestamp' }, { type: 'null' }]
        updated_at: { $ref: '#/components/schemas/Timestamp' }
    UpsertDailyReportRequest:
      type: object
      required: [content, expected_version]
      properties:
        content: { $ref: '#/components/schemas/DailyReportContent' }
        expected_version: { type: integer, minimum: 0 }
    DailyReportPage:
      type: object
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/DailyReport' } }
        next_cursor: { type: [string, 'null'] }
    DailyRewriteRequest:
      type: object
      required: [mode, expected_version]
      properties:
        mode: { type: string, enum: [polish, shorten, structure] }
        expected_version: { type: integer, minimum: 1 }
    ApplyDailyRewriteRequest:
      type: object
      required: [operation_id, content, expected_version]
      properties:
        operation_id: { $ref: '#/components/schemas/Uuid' }
        content: { $ref: '#/components/schemas/DailyReportContent' }
        expected_version: { type: integer, minimum: 1 }
    DepartmentDailyReportItem:
      type: object
      required: [user, work_date, report]
      properties:
        user: { $ref: '#/components/schemas/UserSummary' }
        work_date: { $ref: '#/components/schemas/DateString' }
        report:
          oneOf: [{ $ref: '#/components/schemas/DailyReport' }, { type: 'null' }]
    DepartmentDailyReportView:
      type: object
      required: [department, from, to, items, next_cursor]
      properties:
        department: { $ref: '#/components/schemas/Department' }
        from: { $ref: '#/components/schemas/DateString' }
        to: { $ref: '#/components/schemas/DateString' }
        items: { type: array, items: { $ref: '#/components/schemas/DepartmentDailyReportItem' } }
        next_cursor: { type: [string, 'null'] }